fixes tables

+++ aequilibrae/project/project.py
@@ -91,7 +91,7 @@ class Project:
         self.activate()
 
         self.__create_empty_network()
-        self.__create_empty_transit()
+        self.create_empty_transit()
         self.__load_objects()
         self.about.create()
         global_logger.info(f"Created project on {self.project_base_path}")
@@ -187,7 +187,11 @@ class Project:
         self.conn.commit()
         initialize_tables(self, "network")
 
-    def __create_empty_transit(self):
+    def create_empty_transit(self):
+        """Creates the public transport database"""
+        if not os.path.exists(os.path.join(self.project_base_path, "public_transport.sqlite")):
+            raise FileExistsError("Public Transport database already exists.")
+
         shutil.copyfile(spatialite_database, os.path.join(self.project_base_path, "public_transport.sqlite"))
         self.transit = Transit(self)
         initialize_tables(self, "transit")
